Basic Functionality
This will draw a line in the graph and adds some skeleton code to
organize some of the things that still need to be done.

It's still rough, there's some things that I'd like to see if make the
code cleaner, like a transform matrix.

After parsing formulas decently, I'd like to add tests, adjust the graph
if the formula crosses the y-axis, animate on 1D to see speed, and give
some rendering options for more of a grid, animations, etc.. (also give
a list of common easing functions)